move connector version increment to its own airbyte-ci command
