+++ airbyte-ci/connectors/pipelines/pipelines/airbyte_ci/connectors/test/pipeline.py
@@ -9,11 +9,12 @@ from typing import TYPE_CHECKING
 
 import anyio
 from connector_ops.utils import ConnectorLanguage  # type: ignore
+from pipelines.airbyte_ci.connectors.check_version_increment.pipeline import VersionIncrementCheck
 from pipelines.airbyte_ci.connectors.consts import CONNECTOR_TEST_STEP_ID
 from pipelines.airbyte_ci.connectors.context import ConnectorContext
 from pipelines.airbyte_ci.connectors.reports import ConnectorReport
 from pipelines.airbyte_ci.connectors.test.steps import java_connectors, python_connectors
-from pipelines.airbyte_ci.connectors.test.steps.common import QaChecks, VersionIncrementCheck
+from pipelines.airbyte_ci.connectors.test.steps.common import QaChecks
 from pipelines.helpers.execution.run_steps import StepToRun, run_steps
 
 if TYPE_CHECKING:
